FIX: Collision
Fix collision of the enemy. Added the Win Scene and the door to be able to win. Fix all the warnings, added some sounds with the chest, the door, etc. Fix some volumes. Fix MapManager function (it used the width and height of the map but should be using the tile width and height instead).